Estimate conditional survival with a 95% confidence interval

conditional_surv_est estimates the Kaplan-Meier conditional survival at fixed time points and produces a 95% confidence interval

conditional_surv_est(basekm, t1, t2)

Arguments

basekm

survfit object

t1

the time on which to condition

t2

the survival time to estimate

Value

A list where cs_est is the conditional survival estimate, cs_lci is the lower bound of the 95% confidence interval and cs_uci is the upper bound of the 95% confidence interval

Details

For example, if t1 = 2 and t2 = 5, the function will return the probability of surviving to year 5 conditioned on having already survived to year 2. See the vignette at http://www.emilyzabor.com/condsurv/articles/estimate_cs.html for details on calculations of conditional survival estimates and confidence intervals, and examples.